+++ SecureHttpClient/Platforms/Android/SecureHttpClientHandler.cs
@@ -55,7 +55,7 @@ namespace SecureHttpClient
         /// <param name="pins">The array of certifiate pins (example of pin string: "sha256/fiKY8VhjQRb2voRmVXsqI0xPIREcwOVhpexrplrlqQY=")</param>
         public virtual void AddCertificatePinner(string hostname, string[] pins)
         {
-            _logger?.LogDebug($"Add CertificatePinner: hostname:{hostname}, pins:{string.Join("|", pins)}");
+            _logger?.LogDebug("Add CertificatePinner: hostname:{Hostname}, pins:{Pins}", hostname, string.Join("|", pins));
             CertificatePinning.CertificatePinner.ValidatePattern(hostname); // will throw c# exception instead of java exception if pattern is invalid
             _certificatePinnerBuilder.Value.Add(hostname, pins);
         }